Handle implicit BLOCK_ENTRY arrays
YAML, being what it is, permits both:

    foo:
      - bar
      - baz

and

    foo:
    - bar
    - baz

Unfortunately, libyaml doesn't actually emit an array (mapping) start
token of any kind for the second case.  It just starts emitting
BLOCK_ENTRY tokens.  Since it doesn't emit an *end* token for this,
attmept to detect what I think is the only case this matters (i.e., when
an implicit BLOCK_ENTRY array is inside a dict).

Fun question: what should this do?

    foo:
    - bar
      - bazz

In a better world, it'd be a parse error; in this one, libyaml tokenizes
it into {"foo": ["bar - buzz"]}.

Signed-off-by: Robbie Harwood <rharwood@redhat.com>